+++ packages/flutter_robusta_cli/lib/flutter_robusta_cli.dart
@@ -1,0 +1,4 @@
+/// CLI with abilities to create project and code generation based on Robusta.
+library flutter_robusta_cli;
+
+export 'src/extension.dart';