+++ packages/flutter_robusta_cli/mason_bricks/add_firebase_core/__brick__/lib/robusta/firebase_core.dart
@@ -1,0 +1,3 @@
+import 'package:robusta_firebase_core/robusta_firebase_core.dart';
+
+FirebaseCoreExtension firebaseCoreExtension() => const FirebaseCoreExtension();