+++ packages/flutter_robusta_cli/mason_bricks/new_project/__brick__/lib/main.dart
@@ -1,0 +1,17 @@
+import 'package:flutter_robusta/flutter_robusta.dart';
+import 'package:{{package_name}}/robusta/app.dart';
+import 'package:{{package_name}}/robusta/boot.dart';
+import 'package:{{package_name}}/robusta/event.dart';
+
+Future<void> main() async {
+  final runner = Runner(
+    defineBoot: defineBoot,
+    extensions: [
+      ImplementingCallbackExtension(),
+      appExtension(),
+      eventExtension(),
+    ],
+  );
+
+  await runner.run();
+}