+++ packages/flutter_robusta_cli/mason_bricks/new_project/__brick__/lib/robusta/error_handle.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_robusta/flutter_robusta.dart';
+
+Future<void> onError(ErrorEvent event) async {
+  /// Handle system error,
+  /// you may use rootNavigatorKey to show popup or redirect to error screen?
+}