+++ packages/flutter_robusta_cli/mason_bricks/new_project/__brick__/lib/robusta/event.dart
@@ -1,0 +1,10 @@
+import 'package:flutter_robusta/flutter_robusta.dart';
+import 'package:{{package_name}}/robusta/error_handle.dart';
+
+EventExtension eventExtension() {
+  return EventExtension(
+    configurator: (em, container) {
+      em.addEventListener<ErrorEvent>(onError);
+    },
+  );
+}