+++ packages/robusta_dio/lib/robusta_dio.dart
@@ -1,0 +1,8 @@
+/// Robusta extension for integrating Dio features.
+library robusta_dio;
+
+export 'package:dio/dio.dart';
+
+export 'src/dio.dart';
+export 'src/extension.dart';
+export 'src/provider.dart';