+++ scripts/generate_mason_bricks.dart
@@ -1,0 +1,40 @@
+import 'dart:io';
+
+Future<void> main() async {
+  final stack = {
+    'flutter_robusta_cli': [
+      'new_project',
+      'add_auth',
+      'add_dio',
+      'add_graphql',
+      'add_hive',
+      'add_firebase_core',
+      'gen_repo',
+      'gen_screen',
+    ],
+  };
+
+  for (final entry in stack.entries) {
+    final package = entry.key;
+    final processes = entry.value.map(
+      (brick) => Process.run(
+        'mason',
+        [
+          'bundle',
+          './packages/$package/mason_bricks/$brick',
+          '-t',
+          'dart',
+          '-o',
+          './packages/$package/lib/mason_bricks',
+        ],
+        runInShell: true,
+      ),
+    );
+
+    await Future.wait(processes);
+  }
+
+  await Future.wait([stderr.close(), stdout.close()]);
+
+  exit(0);
+}